# docs/swagger.yaml
definitions:
...
  dto.CreateRoleReq:
    properties:
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    type: object
...
  dto.Permissions:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.RemoveRoleReq:
    properties:
      role_id:
        type: string
    type: object
...
  dto.Role:
    properties:
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/dto.Permissions'
        type: array
    type: object
...
  dto.UpdatePermissionToRoleReq:
    properties:
      permissions:
        items:
          type: string
        type: array
      role_id:
        type: string
    type: object
  dto.UpdatePermissionToRoleRes:
    properties:
      message:
        type: string
      role:
        $ref: '#/definitions/dto.Role'
    type: object
...
  dto.UserRole:
    properties:
      role_id:
        type: string
      user_id:
        type: string
    type: object
  dto.UserRolesRes:
    properties:
      roles:
        items:
          $ref: '#/definitions/dto.Role'
        type: array
      user_id:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        description: Code is the error code. It is not status code
        type: integer
      description:
        description: Description is the error description.
        type: string
      field_error:
        description: FieldError is the error detail for each field, if available that
          is.
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      message:
        description: Message is the error message.
        type: string
      stack_trace:
        description: |-
          StackTrace is the stack trace of the error.
          It is only returned for debugging
        type: string
    type: object
  response.FieldError:
    properties:
      description:
        description: Description is the error description for this field.
        type: string
      name:
        description: Name is the name of the field that caused the error.
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/admin/permissions:
    get:
      consumes:
      - application/json
      description: GetPermissions Retrieve list of permissions.
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      - description: page type (required)
        in: query
        name: page
        required: true
        type: string
      - description: per-page type (required)
        in: query
        name: per-page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Permissions'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: GetPermissions
      tags:
      - Admin
...
  /api/admin/roles:
    delete:
      consumes:
      - application/json
      description: RemoveRole allow user to remove role
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      - description: remove role Request
        in: body
        name: removeReq
        required: true
        schema:
          $ref: '#/definitions/dto.RemoveRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: RemoveRole
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: GetRoles Retrieve list of roles.
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      - description: page type (required)
        in: query
        name: page
        required: true
        type: string
      - description: per-page type (required)
        in: query
        name: per-page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Role'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: GetRoles
      tags:
      - Admin
    patch:
      consumes:
      - application/json
      description: UpdateRolePermissions allow user to update role permissions
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      - description: update role permissions   Request
        in: body
        name: rolePermissions
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePermissionToRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdatePermissionToRoleRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: UpdateRolePermissions
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: CreateRole allow user to create role
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      - description: create role   Request
        in: body
        name: createRoleReq
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Role'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: CreateRole
      tags:
      - Admin
  /api/admin/roles/:id/users:
    get:
      consumes:
      - application/json
      description: GetRoleUsers Retrieve list of users of role.
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Role'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: GetRoleUsers
      tags:
      - Admin
...
  /api/admin/users/:id/roles:
    get:
      consumes:
      - application/json
      description: GetRoleUsers Retrieve list of user roles.
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserRolesRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: GetRoleUsers
      tags:
      - Admin
...
  /api/admin/users/roles:
    post:
      consumes:
      - application/json
      description: RevokeUserRole allow revoke role from user
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      - description: revoke role from user Request
        in: body
        name: userRoleReq
        required: true
        schema:
          $ref: '#/definitions/dto.UserRole'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: RevokeUserRole
      tags:
      - Admin
...
swagger: "2.0"